refactor : 최종 1등을 뽑는 메서드를 전체 데이터를 가진 GameResult 객체가 아니라 ProgressRecord 객체로 이관하여 기록 및 1등 찾기에 간단하게 접근할 수있도록 변경
