+++ src/main/java/racinggame/business/GameMove.java
@@ -7,12 +7,12 @@ import java.util.List;
 
 public record GameMove(NonNegativeMoves moves) {
     
-    public GameResult moveCar(JoinCars joinCars) {
+    public List<ProgressRecord> moveCar(JoinCars joinCars) {
         List<ProgressRecord> progressRecords = new ArrayList<>();
         for(int i = 0; i < this.moves.moves(); i++) {
             joinCars.race(progressRecords);
         }
-        return new GameResult(progressRecords);
+        return progressRecords;
     }
     
 }
