+++ src/main/java/racinggame/business/Process.java
@@ -1,5 +1,6 @@
 package racinggame.business;
 
+import java.util.List;
 import racinggame.model.*;
 import racinggame.response.GameResponse;
 
@@ -7,8 +8,8 @@ public class Process {
     
     public GameResponse run(String[] names, int moves) {
         JoinCars joinCars = CarFactory.createCars(new NonNegativeReadyCars(names));
-        GameResult gameResult = new GameMove(new NonNegativeMoves(moves)).moveCar(joinCars);
-        WinnerCars gameWinners = gameResult.findGameWinners();
-        return new GameResponse(gameResult, gameWinners);
+        List<ProgressRecord> progressRecords = new GameMove(new NonNegativeMoves(moves)).moveCar(joinCars);
+        WinnerCars gameWinners = progressRecords.getLast().findGameWinners();
+        return new GameResponse(new GameResult(progressRecords), gameWinners);
     }
 }
